requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
numpy==1.26.4
python-dotenv==1.0.1
sounddevice==0.4.6
soundfile==0.12.1
supabase==1.2.0
PyJWT==2.8.0
requests==2.31.0